Create SpiralSparse.cpp
This is a program to traverse a sparse matrix (inputted in a row-triplet representation format) in a reverse spiral order starting from the specified position as input by the user.
It traverses in the order down, right, up, left and displays the output in a quadruplet form with 1st column containing row no.,2nd column containing column no.,3rd column containing element and 4th column containing the traversed direction i.e. we display 1 for down,2 for right,3 for up,4 for left.